+++ packages/xrpl/src/models/transactions/transaction.ts
@@ -20,6 +20,9 @@ import { CheckCreate, validateCheckCreate } from './checkCreate'
 import { Clawback, validateClawback } from './clawback'
 import { isIssuedCurrency } from './common'
 import { DepositPreauth, validateDepositPreauth } from './depositPreauth'
+import { DIDDelete, validateDIDDelete } from './DIDDelete'
+import { DIDSet, validateDIDSet } from './DIDSet'
+import { EnableAmendment } from './enableAmendment'
 import { EscrowCancel, validateEscrowCancel } from './escrowCancel'
 import { EscrowCreate, validateEscrowCreate } from './escrowCreate'
 import { EscrowFinish, validateEscrowFinish } from './escrowFinish'
@@ -53,10 +56,12 @@ import {
   PaymentChannelFund,
   validatePaymentChannelFund,
 } from './paymentChannelFund'
+import { SetFee } from './setFee'
 import { SetRegularKey, validateSetRegularKey } from './setRegularKey'
 import { SignerListSet, validateSignerListSet } from './signerListSet'
 import { TicketCreate, validateTicketCreate } from './ticketCreate'
 import { TrustSet, validateTrustSet } from './trustSet'
+import { UNLModify } from './UNLModify'
 import {
   XChainAccountCreateCommit,
   validateXChainAccountCreateCommit,
@@ -88,18 +93,20 @@ import {
  * @category Transaction Models
  */
 export type Transaction =
-  | AccountDelete
-  | AccountSet
   | AMMBid
+  | AMMCreate
   | AMMDelete
   | AMMDeposit
-  | AMMCreate
   | AMMVote
   | AMMWithdraw
+  | AccountDelete
+  | AccountSet
   | CheckCancel
   | CheckCash
   | CheckCreate
   | Clawback
+  | DIDDelete
+  | DIDSet
   | DepositPreauth
   | EscrowCancel
   | EscrowCreate
@@ -119,14 +126,16 @@ export type Transaction =
   | SignerListSet
   | TicketCreate
   | TrustSet
+  | XChainAccountCreateCommit
   | XChainAddAccountCreateAttestation
   | XChainAddClaimAttestation
   | XChainClaim
   | XChainCommit
   | XChainCreateBridge
   | XChainCreateClaimID
-  | XChainAccountCreateCommit
   | XChainModifyBridge
+
+export type PseudoTransaction = EnableAmendment | SetFee | UNLModify
 
 /**
  * @category Transaction Models
@@ -205,16 +214,12 @@ export function validate(transaction: Record<string, unknown>): void {
   // eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- okay here
   setTransactionFlagsToNumber(tx as unknown as Transaction)
   switch (tx.TransactionType) {
-    case 'AccountDelete':
-      validateAccountDelete(tx)
-      break
-
-    case 'AccountSet':
-      validateAccountSet(tx)
-      break
-
     case 'AMMBid':
       validateAMMBid(tx)
+      break
+
+    case 'AMMCreate':
+      validateAMMCreate(tx)
       break
 
     case 'AMMDelete':
@@ -225,16 +230,20 @@ export function validate(transaction: Record<string, unknown>): void {
       validateAMMDeposit(tx)
       break
 
-    case 'AMMCreate':
-      validateAMMCreate(tx)
-      break
-
     case 'AMMVote':
       validateAMMVote(tx)
       break
 
     case 'AMMWithdraw':
       validateAMMWithdraw(tx)
+      break
+
+    case 'AccountDelete':
+      validateAccountDelete(tx)
+      break
+
+    case 'AccountSet':
+      validateAccountSet(tx)
       break
 
     case 'CheckCancel':
@@ -251,6 +260,14 @@ export function validate(transaction: Record<string, unknown>): void {
 
     case 'Clawback':
       validateClawback(tx)
+      break
+
+    case 'DIDDelete':
+      validateDIDDelete(tx)
+      break
+
+    case 'DIDSet':
+      validateDIDSet(tx)
       break
 
     case 'DepositPreauth':
@@ -329,6 +346,10 @@ export function validate(transaction: Record<string, unknown>): void {
       validateTrustSet(tx)
       break
 
+    case 'XChainAccountCreateCommit':
+      validateXChainAccountCreateCommit(tx)
+      break
+
     case 'XChainAddAccountCreateAttestation':
       validateXChainAddAccountCreateAttestation(tx)
       break
@@ -351,10 +372,6 @@ export function validate(transaction: Record<string, unknown>): void {
 
     case 'XChainCreateClaimID':
       validateXChainCreateClaimID(tx)
-      break
-
-    case 'XChainAccountCreateCommit':
-      validateXChainAccountCreateCommit(tx)
       break
 
     case 'XChainModifyBridge':